machines: moby: enable impermanence (experimental)
last time i tried to rebuild nixos on the pinephone it didn't switch,
IIRC. i don't know why, but i expect it'll be easier to manage my
machines if i keep more of the setup consistent across them.

likely i'll tune this to be more leniant on moby (e.g. persist all of
.librewolf).